requirements.txt
# core
django>=3.2.14
djangorestframework

# style
black
isort
